Some Python 3 fixes

+++ gcloudc/db/backends/datastore/utils.py
@@ -465,6 +465,6 @@ def ensure_datetime(value):
         Painfully, sometimes the Datastore returns dates as datetime objects, and sometimes
         it returns them as unix timestamps in microseconds!!
     """
-    if isinstance(value, long):
+    if isinstance(value, int):
         return datetime.fromtimestamp(value / 1e6)
     return value
